Python-Script/02-Simple-Monitoring-Dashboard/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.25